chore(*) separate the resolver in a separate module for DSN
