+++ pkg/dns-server/resolver/handler.go
@@ -1,4 +1,4 @@
-package dns_server
+package resolver
 
 import (
 	"fmt"
