+++ pkg/dns-server/resolver/ip.go
@@ -1,4 +1,4 @@
-package dns_server
+package resolver
 
 import (
 	"net"
