+++ pkg/dns-server/resolver/ip_test.go
@@ -1,4 +1,4 @@
-package dns_server
+package resolver
 
 import (
 	"math"
