+++ pkg/dns-server/resolver/server_test.go
@@ -1,4 +1,4 @@
-package dns_server
+package resolver
 
 import (
 	"fmt"
