+++ pkg/dns-server/resolver/utils_test.go
@@ -1,4 +1,4 @@
-package dns_server
+package resolver
 
 import (
 	. "github.com/onsi/ginkgo"
