+++ pkg/dns-server/synchroniser/sycnroniser.go
@@ -1,0 +1,1 @@
+package synchroniser
